requirements.txt
fastapi[all]
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
sqlalchemy
psycopg2-binary